requirements.txt
aiohttp~=3.7.3
aiohttp-jinja2
jinja2~=2.11.3



pyserial~=3.5
flask~=1.1.2